Turn diceware module into a package.
Move wordlists into the new package. This should fix
the data_files problem on most platforms and in most
environments.

Seems to be really tricky to install data files in a
foreseeable location.

The more finegrained solution of conorsch might be
slightly less stable in the long run, therefore we
selected the 'package', actually.