Remove `diceware` run in travis.
The `diceware` run in travis was meant as a packaging test. It was
supposed to fail if not all data files were available in
installs. Unfortunately it did not fail in cases where `pip` *did*
fail.

Furthermore we want to detect such packaging problems in regular
tests, before pushing changes.