Support pluggable sources of randomness.
Based on suggestions of @drebs we now start to support different
sources of randomness. We start with one source, the default,
which will be the standard libs `random.SystemRandom` source
we used anyway. But from now on this source if put into a
separate class, tested and all that, which will be registered
as an entry point.

Yet we do not register or work with the registered sources. This
will come next.