Start RealDiceRandomSource.
This will only be a skeleton, to be filled later.

We got an `input_func` which works regardless of Python version (as `raw_input`
was renamed `input` with Python 3.

We also got a RealDiceRandomSource class without any functionality
yet. The `get_input` method is only a temporary placeholder to enable
testing of mocking the input_func.