Rename default wordlist from "en_8k" to "en".
Normally, it would be more concise to use the more declarative name.
"en_8k" tells, how many terms we find in this wordlist. But as the
wordlist names are used in the CLI and users will normally simply
look for a wordlist in their language, it can help to keep the name
of the "default" list for each language short.

We pick the 8K lists for "default" (and not the "original" 6^5
lists), because we assume that normally people will use diceware
without dice. Also most other sources of randomness will most
probably work on a 2^x basis rather than in 6^x (and therefore
have better use for 8k lists over other ones).