Add modulo functionality
Add an if to the pre_check that catches cases where there are less items
to choose from than sides on the die. If this is the case, we can still
pick items, but we have to use a modulo.
Pick the `num_rolls` so that the result generated after rolling is larger
 than `100*len(sequence)`.
The `pre_check()` function now has to return the modified `num_rolls` as
well as a boolean to indicate whether modulo needs to be applied on the
result.