Remove overcomplicated stuff from randomsource.
While it is nice to have a working plugin example, it looks also a bit
overcomplicated to me. It is better for contributors, if they can find
things (like CLI argument parsing) in one (foreseeable) place and do
not have for things scattered over the whole project.

For plugin documentation we can use the sphinx docs. Or maybe provide
a sample plugin in tests.