Move pytest config from `setup.cfg` to `tox.ini`.
Use of ``[pytest]`` section in `setup.cfg` is deprecated while
the new section name ``[tool:pytest]`` is not recognized by older
`pytest` versions.

To stay compatible with older `pytest` and avoid warnings with
`pytest` 3.0+, we move the whole config to `tox.ini`, where still
the old section name is valid.