Disable packaging-related tests by default.
We now `pytest.mark` tests that are mainly related to packaging and
disable them by default. This is a first workaround for broken
year-checks. These tests blindly fail when the current year does not
equal the one mentioned in docs and version string.

In fact the tests should compare the year in docs/version string with
the year of "current" commit (with "current commit" being the one
representing the tested codebase). Another issue to be addressed.

All tests can be run by: `py.test -m ""`, the packaging-only tests can
be run by `py.test -m "packaging"`. `tox` got a new env called `pkg`.