Remove discriminating terms (fixes #85)
Some wordlists contained vulgar, discriminating, and in in other
respects bad language. We start to fix that by removing/replacing terms
found by @kmille. But there are certainly more words to remove and we
must face the problem, that it is not easy to generate proper welcoming
wordlists automatically. Any further hints/help welcome.