Use `pyproject.toml` for dependencies of docs.
As tools to generate docs are now defined in `pyproject.toml`, we use
them to generate a list of pinned dependencies (as recommended by
rtfm.org).

This way we still have to update the dependency list
`docs/requirements.txt` manually from time to time, but now we only use
a single source of base dependencies - the `docs`-extra dependency of
the package.